feat(app): handle http errors (#54)
This refactors the `ExceptionService` to implement the `ErrorHandler`
interface and registers it as the application-wide error handler. The
`ExceptionService` now handles all errors that are thrown or not handled
by other error handlers. Errors are currently handled by logging them to
the console. We may want to send errors to a monitoring service here and
open a snackbar or dialog for appropriate errors messages to inform the
user.

An new method `ExceptionService.handleHttpError` that returns an
`OperatorFunction` piping its source observable through the `catchError`
allows for handling of `HttpErrorResponses`.

A new interceptor `HandleHttpErrorInterceptor` has been created and
registered as a multi `HTTP_INTERCEPTORS` provider. This interceptor
pipes the source http event stream through the operator function
returned from the `ExceptionService.handleHttpError` method. This allows
for all http errors to be handled without having to add the `catchError`
operator to every http request stream.