fix(map): fix multi dwelling color when filteres are applied
When filters are aplied to the building data the avg ECP color for multi
dwelling builfings is updated to reflect the new avage of filtered data
as expected. However if a filter that caused a multi dwelling building
to include only a single dwelling the color would be updated to a solid
EPC color causing confusion to the user.

This fixes the issues by checking if the building was originally a multi
dwelling building by checking the unfiltered building data. If so the
avg EPC pattern is applied to the building not the solid EPC color, to
indicate that the building is a multi dwelling building.